vfs_fruit: fix fruit:resource option spelling, but not behaviour
IMPORTANT
=========

this is a backport of commit 7d4b4a429955cf7fd43f3ac83990c5386ac8cd76
that only uses the spelling corrections of comments and the tests
without changing any behaviour. This greatly simplifies the backport.

Changing the option names in the tests without fixing the code means
that this patchset can't be used to run autobuild...

ORIGINAL COMMIT MESSAGE
=======================

Fix all occurences of bad spelling of "resource" as "res*s*ource" (two
s).

One of the places where this was wrong was when parsing parametric
options in the VFS connect() function in the module. As a result any
setting of

  fruit:resource=something

in smb.conf was silently ignored and the default ("file") was active.

In Samba 4.6 we accept both the wrong and the correct spelling, in Samba
4.7 the bad spelling will be removed.

BUG: https://bugzilla.samba.org/show_bug.cgi?id=12412

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Jeremy Allison <jra@samba.org>
(adopted from commit 7d4b4a429955cf7fd43f3ac83990c5386ac8cd76)